% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ai}
\alias{ai}
\title{AI association data}
\format{
A \code{tibble} of 2,500 AI-associations with the following columns:
\describe{
 \item{id}{Participant id}
 \item{trial}{Index of response. Every participant produced five associations.}
 \item{text}{Text of responses.}
}
}
\usage{
ai
}
\description{
Excerpt data from Scharowski, Mata, and Wulff (2024) containing laypeople associations of artificial intelligence.
}
\references{
Scharowski, S., Mata, R., Wulff, D. U. (2024). Laypeople's associations with AI. In preparation.
}
\keyword{datasets}
