% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Cluster embedding or projection vectors}
\usage{
cluster(
  embedding,
  method = "hclust",
  k = NULL,
  eps = NULL,
  metric = "arccos",
  ...,
  verbose = FALSE
)
}
\arguments{
\item{embedding}{a \code{numeric} matrix containing a text embedding or embedding projection.}

\item{method}{a \code{character} string specifying the clustering method One of \code{c("hclust","dbscan","louvain")}. Default is \code{"hclust"}.}

\item{k}{an \code{integer} specifying the number of clusters for \code{method = "hclust"}.}

\item{eps}{a \code{numeric} specifying the point distance within clusters for \code{method = "dbscan"}.}

\item{metric}{a \code{character} string specifying the similarity function used for methods \code{c("hclust","louvain")}.}

\item{...}{further arguments passed on to the clustering methods.}

\item{verbose}{a \code{logical} specifying whether to show messages.'}
}
\value{
The function returns a \code{matrix} containing the input embedding or projection and a new attribute \code{"cluster"}.
}
\description{
\code{cluster} generates clusters of embedding vectors using standard clustering algorithms.
}
\examples{

# add clustering to embedding
projection <- embed(neo$text) \%>\%
  cluster()

}
\references{
Wulff, D. U., Aeschbach, S., & Mata, R. (2024). embeddeR. psyArXiv
}
