% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cmix}
\alias{cmix}
\title{Print embedding object}
\usage{
cmix(col1, col2, weight, format = "hex")
}
\arguments{
\item{col1, col2}{color values as label, rgb or hex.}

\item{weight}{a numeric value specifying the amount of weight given to \code{col1} and \code{col2}}

\item{format}{a character string specifying the output type. One of \code{c("rgb","hex")}. Default is \code{"hex"}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Generic function \code{print.embedR} shows a print of the embedding object with an abbreviated \code{matrix} and a \code{tibble} of its attributes.
}
