% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_media.R
\name{embed_audio}
\alias{embed_audio}
\title{Embed audio in R Markdown documents}
\usage{
embed_audio(
  src,
  type = c("mpeg", "ogg", "wav"),
  attribute = c("controls", "autoplay", "loop", "muted", "preload"),
  placeholder = ""
)
}
\arguments{
\item{src}{A path or URL to the media file.}

\item{type}{The type of media file specified in \code{src}.}

\item{attribute}{A character vector specifying which attributes to use. An
empty character vector "" can be used if no attributes are desired.}

\item{placeholder}{The placeholder text to use when the output format is
not HTML.}
}
\value{
If \code{knitr::is_html_output()} == \code{TRUE}, returns HTML \verb{<video>} code.
If \code{knitr::is_html_output()} = \code{FALSE}, returns placeholder text.
}
\description{
\code{embed_audio()} provides a standard way to embed audio in R Markdown
documents when the output format is HTML, and to print placeholder text
when the output format is not HTML.
}
\details{
\code{embed_audio()} is a wrapper for the HTML5 \verb{<audio>} element that prints
HTML \verb{<audio>} code in HTML documents built by R Markdown and placeholder
text in non-HTML documents built by R Markdown. This function may be useful
for conditional output that depends on the output format. For example, you
may embed audio in an R Markdown document when the output format is HTML,
and print placeholder text when the output format is LaTeX.

The function determines output format using \code{\link[knitr:is_html_output]{knitr::is_html_output()}}. By
default, these formats are considered as HTML formats: \code{c('markdown', 'epub', 'html', 'html5', 'revealjs', 's5', 'slideous', 'slidy')}.
}
\note{
This function is supposed to be used in R code chunks or inline R code
expressions. You are recommended to use forward slashes (/) as path
separators instead of backslashes in the file paths.
}
