% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_media.R
\name{embed_video}
\alias{embed_video}
\title{Embed video in R Markdown documents}
\usage{
embed_video(
  src,
  type = c("mp4", "webm", "ogg"),
  width = "320",
  height = "240",
  attribute = c("controls", "autoplay", "loop", "muted", "preload", "none"),
  thumbnail = NULL,
  id = "",
  placeholder = ""
)
}
\arguments{
\item{src}{A path or URL to the media file.}

\item{type}{The type of media file specified in \code{src}.}

\item{width}{The width of the video, in pixels.}

\item{height}{The height of the video, in pixels.}

\item{attribute}{A character vector specifying which attributes to use.
"none" can be used if no attributes are desired.}

\item{thumbnail}{A path to an image.}

\item{id}{A character string specifying a unique ID for the element.
Can be used by CSS or JavaScript to perform certain tasks for
the element with the specific ID.}

\item{placeholder}{The placeholder text to use when the output format is
not HTML.}
}
\value{
If \code{knitr::is_html_output()} is \code{TRUE}, returns HTML \verb{<video>} code.
If \code{knitr::is_html_output()} is \code{FALSE}, returns placeholder text.
}
\description{
\code{embed_video()} provides a standard way to embed video in R Markdown
documents when the output format is HTML, and to print placeholder text
when the output format is not HTML.
}
\details{
\code{embed_video()} is a wrapper for the HTML5 \verb{<video>} element that prints
HTML \verb{<video>} code in HTML documents built by R Markdown and placeholder
text in non-HTML documents built by R Markdown. This function may be useful
for conditional output that depends on the output format. For example, you
may embed video in an R Markdown document when the output format is HTML,
and print placeholder text when the output format is LaTeX.

The function determines output format using \code{\link[knitr:is_html_output]{knitr::is_html_output()}}. By
default, these formats are considered as HTML formats: \code{c('markdown', 'epub', 'html', 'html5', 'revealjs', 's5', 'slideous', 'slidy')}.
}
\note{
This function is supposed to be used in R code chunks or inline R code
expressions. You are recommended to use forward slashes (/) as path
separators instead of backslashes in the file paths.
}
