% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{er_cluster}
\alias{er_cluster}
\title{Cluster embedding vectors}
\usage{
er_cluster(
  embedding,
  method = "hclust",
  k = NULL,
  eps = NULL,
  metric = "arccos",
  ...,
  verbose = FALSE
)
}
\arguments{
\item{embedding}{a \code{numeric} matrix containing a text embedding.}

\item{method}{a \code{character} string specifying the clustering method One of \code{c("hclust","dbscan","louvain")}. Default is \code{"hclust"}.}

\item{k}{an \code{integer} specifying the number of clusters for \code{method = "hclust"}.}

\item{eps}{a \code{numeric} specifying the within-cluster point distance for \code{method = "dbscan"}.}

\item{metric}{a \code{character} string specifying the similarity function used for methods \code{c("hclust","louvain")}.}

\item{...}{further arguments passed on to the clustering methods. Can be used, e.g., to specify the linkage criterion in hierarchical clustering (see \link[stats]{hclust}), the minimum number of points in DBSCAN clustering (see \link[dbscan]{dbscan}), or the resolution in Louvain clustering (see \link[igraph]{cluster_louvain}).}

\item{verbose}{a \code{logical} specifying whether to show messages.'}
}
\value{
The function returns a \code{matrix} containing the input embedding, which has gained a new attribute \code{"cluster"}.
}
\description{
Function \code{er_cluster} generates clusters of embedding vectors using standard clustering algorithms.
}
\examples{

# add clustering to embedding
embedding <- er_cluster(embedding)

}
\references{
Wulff, D. U., Aeschbach, S., Hussain, Z., & Mata, R. (2024). embeddeR. In preparation.
}
