% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{er_compare_embeddings}
\alias{er_compare_embeddings}
\title{Compare embeddings}
\usage{
er_compare_embeddings(
  embeddings,
  metric = "cosine",
  comparison_metric = "spearman"
)
}
\arguments{
\item{embeddings}{a \code{list} of embedding matrices.}

\item{metric}{a \code{character} string specifying the type of similarity used to produce similarity matrices from each embedding. One of \code{c("cosine","arccos","pearson","spearman")}. Default is \code{"cosine"}.}

\item{comparison_metric}{a \code{character} string specifying the type of similarity used to compare similarity matrices. One of \code{c("cosine","arccos","pearson","spearman")}. Default is \code{"spearman"}.}
}
\value{
The function returns a \code{tibble} containing for every pair of embeddings the representational similarities overall (based on the lower triangle) and per row (text).
}
\description{
Function \code{er_compare_embeddings} computes the similarity between two or more embeddings.
}
\examples{
\dontrun{
# get embedding
embedding_1 <- embed(neo$text)
embedding_2 <- embed(neo$text, model = "distilbert-base-uncased")

# compute similarity
compare_embeddings(list(embedding_1, embedding_2))
}
}
\references{
Wulff, D. U., Aeschbach, S., Hussain, Z., & Mata, R. (2024). embeddeR. In preparation.
}
