% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{er_compare_vectors}
\alias{er_compare_vectors}
\title{Compare embedding vectors}
\usage{
er_compare_vectors(embedding, metric = "cosine")
}
\arguments{
\item{embedding}{a \code{numeric} matrix containing a text embedding.}

\item{metric}{a \code{character} string specifying the type of similarity. One of \code{c("cosine","arccos","pearson","spearman")}. Default is \code{"cosine"}.}
}
\value{
The function returns a \code{matrix} containing the similarities of all pairs of embedding vectors. The \code{matrix} has \code{nrow(embedding)} rows and columns.
}
\description{
Function \code{er_compare_vectors} computes similarities of embedding vectors.
}
\examples{

# compute similarity matrix
embedding <- compare_vectors(embedding)

}
\references{
Wulff, D. U., Aeschbach, Hussain, Z., S., & Mata, R. (2024). embeddeR. In preparation.
}
