% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{er_embed}
\alias{er_embed}
\title{Generate text embeddings}
\usage{
er_embed(text, api = "huggingface", model = NULL, type = NULL, verbose = FALSE)
}
\arguments{
\item{text}{a \code{character} vector of texts.}

\item{api}{a \code{character} string specifying the embedding API. One of \code{c("huggingface","openai","cohere")}. Default is \code{"huggingface"}.}

\item{model}{a \code{character} string specifying the embedding model. Must match the model names in the corresponding APIs. See, \href{https://huggingface.co/models}{huggingface.co/models}, \href{https://platform.openai.com/docs/models/embeddings}{platform.openai.com/docs/models/embeddings}, \href{https://cohere.com/embeddings}{cohere.com/embeddings}. Defaults to \code{"sentence-transformers/all-mpnet-base-v2"} for \code{api = "huggingface"}, to \code{"text-embedding-ada-002"} for \code{api = "openai"}, and to \code{"embed-english-v3.0"} for \code{api = "cohere"}.}

\item{type}{a \code{character} string specifying the type of Cohere embeddings. One of \code{c("search_document","search_query","classification","clustering")}. Default is \code{"clustering"}. See \href{https://docs.cohere.com/reference/embed}{https://docs.cohere.com/reference/embed}.}

\item{verbose}{a \code{logical} specifying whether to show messages.}
}
\value{
The function returns a \code{matrix} containing the text embeddings with \code{length(text)} rows and as many columns as there are embedding dimensions.
}
\description{
Function \code{er_embed} generates embeddings for text inputs.
}
\examples{
#'\dontrun{
# run embedding
er_embed(neo$text)
}
}
\references{
Wulff, D. U., Aeschbach, S., Hussain, Z., & Mata, R. (2024). embeddeR. In preparation.
}
