% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{er_get_tokens}
\alias{er_get_tokens}
\title{Get API tokens}
\usage{
er_get_tokens()
}
\value{
The function returns a \code{tibble} showing previously defined access tokens.
}
\description{
Function \code{er_get_tokens} lists existing access tokens of embedding and inference APIs.
}
\examples{

# retrieve access tokens
er_get_tokens()

}
\references{
Wulff, D. U., Aeschbach, S., Hussain, Z., & Mata, R. (2024). embeddeR. In preparation.
}
