% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{er_infer_labels}
\alias{er_infer_labels}
\title{Infer category labels}
\usage{
er_infer_labels(
  labels,
  api = "huggingface",
  model = NULL,
  role = "assistant",
  instruct = NULL,
  system = NULL,
  verbose = FALSE
)
}
\arguments{
\item{labels}{a \code{list} of character vectors.}

\item{api}{a \code{character} string specifying the api One of \code{c("huggingface","openai","cohere")}. Default is \code{"openai"}.}

\item{model}{a \code{character} string specifying the model label. Must match the model names on the corresponding APIs. See, \href{https://huggingface.co/models}{huggingface.co/models} and \href{https://platform.openai.com/docs/models/embeddings}{platform.openai.com/docs/models/embeddings}. Defaults to \code{"meta-llama/Llama-2-70b-chat-hf"} for \code{api = "huggingface"} and to \code{"gpt-4"} for \code{api = "openai"}.}

\item{role}{a \code{character} string specifying the systems role in place of \code{role} in the general system instruction to the model. Default is \code{"assistant"}.}

\item{instruct}{a \code{character} string specifying the instruction for the model. Must contain the placeholder \code{"{examples}"}. Default is \code{"Generate a specific and accurate one or two word category label that captures the common meaning of the following examples: {examples}. Place '@' before and after the category label."}.}

\item{system}{a \code{character} string specifying the general system instruction to the model. Default is \code{"You are a helpful {role} who provides short, specific, and accurate category labels."}.}

\item{verbose}{a \code{logical} specifying whether to show messages.'}
}
\value{
The function returns a \code{character} vector of category labels.
}
\description{
\code{er_infer_labels} infers category labels using generative large language models.
}
\details{
The models recommended for label inferences, including the default models, are not free for use and using them can result in significant costs. Costs will depend on the size of input texts and the number of labels inferred. The default Hugging Face model, \code{meta-llama/Llama-2-70b-chat-hf}, requires a PRO subscription at a monthly price. The OpenAI models, including the default \code{gpt-4} model, incur costs based on the number of tokens in the input and output.

To obtain the best possible labels it is recommended to adjust the prompt arguments \code{role}, \code{system}, and \code{instruct}.
}
\examples{

# get labeled results
result <- er_embed(neo$text) \%>\%
  er_group() \%>\%
  er_project() \%>\%
  er_frame() \%>\%
  dplyr::mutate(group_labels = er_infer_labels(group_texts))

}
\references{
Wulff, D. U., Aeschbach, S., Hussain, Z., & Mata, R. (2024). embeddeR. psyArXiv
}
