% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{er_project}
\alias{er_project}
\title{Project embedding}
\usage{
er_project(embedding, method = "mds", k = 2, ..., verbose = FALSE)
}
\arguments{
\item{embedding}{a \code{numeric} matrix containing a text embedding.}

\item{method}{a \code{character} string specifying the type of projection. One of \code{c("mds","umap","pacmap")}. Default is \code{"mds"}. Projection method \code{"PaCMAP"} is based on the Python library \code{pacmap} (see \href{https://pypi.org/project/pacmap/}{pypi.org/project/pacmap/}) and requires access to a python environment}

\item{k}{an \code{integer} determining the number of dimensions. Default is \code{2}.}

\item{...}{additional parameters handed to the projection method.}

\item{verbose}{a \code{logical} indicating whether to show messages.}
}
\value{
The function returns a \code{matrix} containing an embedding with columns representing the projected coordinates. The \code{matrix} has \code{nrow(embedding)} rows and \code{k} columns.
}
\description{
Function \code{er_project} projects the embedding into a two or more dimensional space using the non-linear dimensionality reduction techniques MDS, UMAP, and PaCMAP.
}
\examples{
\dontrun{
# project embedding
embedding <- embedding \%>\%
  er_project()
}
}
\references{
Wulff, D. U., Aeschbach, S., Hussain, Z., & Mata, R. (2024). embeddeR. In preparation.
}
