% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{er_set_tokens}
\alias{er_set_tokens}
\title{Set API tokens}
\usage{
er_set_tokens(..., hard = FALSE)
}
\arguments{
\item{...}{one or more api-token (name-value) pairs (e.g., huggingface = "TOKEN"). Names can be one or more of \code{c("huggingface","openai","cohere")}.}

\item{hard}{a \code{logical} specifying whether existing tokens should be overwritten.}
}
\description{
Function \code{er_set_tokens} sets access tokens of embedding and inference APIs.
}
\details{
To obtain API tokens users must register with the respective services:

\itemize{
 \item{"Hugging Face"}{Sign up at \href{https://huggingface.co/inference-api}{huggingface.co/inference-api}. Then go to profile and click on settings. Generate access token in the \emph{Access Tokens} sub-menu.}
 \item{"OpenAI"}{Sign up at \href{https://openai.com/blog/openai-api}{openai.com/blog/openai-api}. Select API. Click on Personal (top-right) and select \emph{View API keys}. Then \emph{Create new secret key}.}
 \item{"Cohere"}{Sign up at \href{https://cohere.com/}{cohere.com}. Select \emph{API keys} in your personal dashboard. Either \emph{Create Trial key} (free of charge, but rate limited) or \emph{Create Production key} (costly, but unlimited). }
}

\emph{Hugging Face} and \emph{Cohere} both offer free-of-charge rate-limited API use.
}
\examples{
\dontrun{
# set hugging face token
set_tokens("huggingface" = "TOKEN")
}
}
\references{
Wulff, D. U., Aeschbach, S., Hussain, Z., & Mata, R. (2024). embeddeR. In preparation.
}
