% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{frame}
\alias{frame}
\title{Create data frame}
\usage{
frame(embedding)
}
\arguments{
\item{embedding}{a \code{numeric} matrix containing a text embedding projection.}
}
\value{
The function returns a \code{tibble} containing a projected embedding. The projection dimensions will be assigned to columns \code{x} and \code{y}.
}
\description{
\code{frame} transforms embedR object to a data frame of class \code{tibble}.
}
\examples{

# embed, project, and frame
tbl <- embed(neo$text) \%>\%
  project() \%>\%
  frame()

}
\references{
Wulff, D. U., Aeschbach, S., & Mata, R. (2024). embeddeR. psyArXiv
}
