% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{get_tokens}
\alias{get_tokens}
\title{Get API tokens}
\usage{
get_tokens()
}
\value{
The function returns a \code{tibble} showing existing access tokens.
}
\description{
\code{get_tokens} lists existing access tokens of embedding APIs.
}
\examples{

# get embedding
embedding <- embed(neo$text)

# compute similarity
compare_vectors(embedding)

}
\references{
Wulff, D. U., Aeschbach, S., & Mata, R. (2024). embeddeR. psyArXiv
}
