% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group}
\alias{group}
\title{Group embedding or projection}
\usage{
group(embedding, method = "identity", threshold = 0.995, verbose = FALSE)
}
\arguments{
\item{embedding}{a \code{numeric} matrix containing a text embedding or embedding projection.}

\item{method}{a \code{character} string specifying the reduction method One of \code{c("identity","fuzzy")}. Default is \code{"identity"}.}

\item{threshold}{a \code{numeric} specifying the threshold for \code{method = fuzzy}.}

\item{verbose}{a \code{logical} specifying whether to show messages.'}
}
\value{
The function returns a \code{matrix} containing the embedding or projection. The \code{matrix} has \code{ncol(embedding)} dimensions and as many rows as determined by the condensing method. The \code{matrix} will gain the attribute \code{counts} containing the frequency table of each element in the original \code{embedding} or \code{projection}.
}
\description{
\code{group} transforms an embedding projections to a tibble ready for \code{ggplot2}.
}
\examples{

# get embedding
projection <- embed(neo$text) \%>\%
  project() \%>\%
  to_tibble()

}
\references{
Wulff, D. U., Aeschbach, S., & Mata, R. (2024). embeddeR. psyArXiv
}
