% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media-files.R
\docType{data}
\name{media-files}
\alias{media-files}
\alias{mp3}
\alias{mp4}
\alias{png}
\title{Example Media Files}
\format{
An object of class \code{character} of length 1.
}
\usage{
mp3
}
\description{
Example media files included with embedr.
}
\details{
\itemize{
\item \code{mp3}: MP3 audio
\item \code{mp4}: MP4 video
\item \code{png}: PNG thumbnail
}
}
\keyword{datasets}
