% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neo}
\alias{neo}
\title{NEO personality items}
\format{
A \code{tibble} of 300 personality items with the following columns:
\describe{
 \item{instrument}{Abbreviated name of personality questionnaire.}
 \item{label}{Scale label. Scales each aggregate ten items}
 \item{key}{Direction of item. Used in scale construction.}
 \item{text}{Text of personality item.}
}
}
\source{
\url{https://ipip.ori.org/}
}
\usage{
neo
}
\description{
Personality items from the 300-item NEO personality questionnaire obtained from \href{https://ipip.ori.org/}{IPIP}.
}
\references{
Wulff, D. U., & Mata, R. (2023, October 12). Automated jingle–jangle detection: Using embeddings to tackle taxonomic incommensurability. https://doi.org/10.31234/osf.io/9h7aw
}
\keyword{datasets}
