% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{pacmap}
\alias{pacmap}
\title{Run PaCMAP}
\usage{
pacmap(
  embedding,
  n_components = 2,
  n_neighbors = 10,
  MN_ratio = 0.5,
  FP_ratio = 2,
  distance = "euclidean",
  lr = 1,
  num_iters = 450,
  verbose = FALSE,
  apply_pca = TRUE
)
}
\arguments{
\item{embedding}{a \code{numeric} matrix containing a text embedding.}

\item{n_components}{an \code{integer} Dimensions of the embedded space. Default is \code{2}.}

\item{n_neighbors}{an \code{integer} specifying the number of neighbors considered for nearest neighbor pairs for local structure preservation. Default is \code{10}.}

\item{MN_ratio}{a \code{numeric} specifying the ratio of mid-near pairs to nearest-neighbor pairs  (e.g. \code{n_neighbors=10}, \code{MN_ratio=0.5} means 5 mid-near pairs). Mid-near pairs are used for global structure preservation. Default is \code{.5}.}

\item{FP_ratio}{a \code{numeric} specifying the ratio of further pairs to nearest-neighbor pairs (e.g. \code{n_neighbors=10}, \code{FP_ratio=2} means 20 further pairs). Further pairs are used for both local and global structure preservation. Default is \code{2}.}

\item{distance}{a \code{character} string specifying the distance metric. One of \code{c("euclidean", "manhattan", "angular", "hamming")}. Default is \code{"euclidean"}.}

\item{lr}{a \code{numeric} specifying the learning rate of the Adam optimizer for embedding. Default is \code{1}.}

\item{num_iters}{an \code{integer} specifying the number of iterations for the optimization of embedding. Values greater than 250 are recommended. Default is \code{450}.}

\item{verbose}{a \code{logical} specifying whether to show messages during initialization and fitting. Default is \code{FALSE}.}

\item{apply_pca:}{a \code{logical} specifying whether to apply PCA on the data before pair construction. Default is \code{FALSE}.}
}
\value{
The function returns a \code{matrix} containing projected coordinates for each embedding vectors. The \code{matrix} has \code{nrow(embedding)} rows and \code{n_components} columns.
}
\description{
Run PaCMAP dimensionality reduction
}
\details{
Function wraps around the PaCMAP Python module found at \href{https://github.com/YingfanWang/PaCMAP}{github.com/YingfanWang/PaCMAP}. Function adapted from \href{https://github.com/milescsmith/ReductionWrappers}{/github.com/milescsmith/ReductionWrappers}.

PaCMAP (Pairwise Controlled Manifold Approximation) Maps high-dimensionaldataset to a low-dimensional embedding. For details see \href{https://www.jmlr.org/papers/volume22/20-1061/20-1061.pdf}{jmlr.org/papers/volume22/20-1061/20-1061.pdf}.
}
