% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.embedR_tbl}
\alias{plot.embedR_tbl}
\title{Plot embedding frame}
\usage{
\method{plot}{embedR_tbl}(
  data,
  dimensions = c(1, 2),
  size = NULL,
  color = NULL,
  label = NULL,
  label_size = 3,
  label_filter = NULL,
  pt_padding = 0.05,
  box_padding = 0.05,
  viridis_set = "E",
  viridis_limits = c(0, 0.9)
)
}
\arguments{
\item{data}{a \code{data.frame} created using \link[emebedR]{er_frame}.}

\item{dimensions}{an \code{integer} vector determining the dimensions plotted. Default is \code{c(1, 2)}.}

\item{size}{an optional \code{character} string specifying the column used to determine point size.}

\item{color}{an optional \code{character} string specifying the column used to determine point (and label) color.}

\item{label}{an optional \code{character} string specifying the column used to determine label texts.}

\item{label_size}{an optional \code{character} string specifying the column used to determine label size.}

\item{label_filter}{an optional \code{logical} comparison determining which label texts to show in the plot. Can be based on any columns in \code{data}.}

\item{pt_padding}{a \code{numeric} specifying the point padding.}

\item{box_padding}{a \code{numeric} specifying the label box padding.}

\item{viridis_set}{a \code{character} string determining the viridis color set. See \link[ggplot2]{scale_color_viridis_d}.}

\item{viridis_limits}{a \code{numeric} vector of length 2 determining the \code{begin} and \code{end} arguments of \link[ggplot2]{scale_color_viridis_d}.}
}
\value{
The function returns a \code{ggplot2} object.
}
\description{
Generic function \code{plot} shows the points in a two-dimensional embedding or projection space with options for customization.
}
\examples{

# get embedding plot
neo$text \%>\%
  er_embed() \%>\%
  er_project() \%>\%
  er_frame()
  plot()

}
\references{
Wulff, D. U., Aeschbach, S., Hussain, Z., & Mata, R. (2024). embeddeR. In preparation.
}
