% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Plot projection}
\usage{
plot(embedding, method = "mds", k = 2, ..., verbose = FALSE)
}
\arguments{
\item{embedding}{a \code{numeric} matrix containing a text embedding.}

\item{method}{a \code{character} string specifying the type of projection. One of \code{c("mds","umap","pacmap")}. Default is \code{"mds"}.}

\item{k}{an \code{integer} determining the number of dimensions. Default is \code{2}.}

\item{...}{additional parameters handed to the embedding method.}

\item{verbose}{a \code{logical} indicating whether to show messages.}
}
\value{
The function returns a \code{matrix} containing projected coordinates for each embedding vectors. The \code{matrix} has \code{nrow(embedding)} rows and \code{k} columns.
}
\description{
\code{project} projects the embedding into a two or more dimensional space.
}
\examples{

# get embedding
embedding <- embed(neo$text)

# project embedding
project(embedding)

}
\references{
Wulff, D. U., Aeschbach, S., & Mata, R. (2024). embeddeR. psyArXiv
}
