% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.embedR}
\alias{print.embedR}
\title{Print embedding object}
\usage{
\method{print}{embedR}(x, n = 5, m = 5, ...)
}
\arguments{
\item{x}{an object of class embedR}

\item{n}{an \code{integer} specifying the number of rows to print. Default is \code{5}.}

\item{m}{an \code{integer} specifying the number of columns to print. Default is \code{5}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Generic function \code{print.embedR} shows a print of the embedding object with an abbreviated \code{matrix} and a \code{tibble} of its attributes.
}
