% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{set_tokens}
\alias{set_tokens}
\title{Set API tokens}
\usage{
set_tokens(..., hard = FALSE)
}
\arguments{
\item{...}{one or more api-token (name-value) pairs (e.g., huggingface = "TOKEN"). Names can be one or more of \code{c("huggingface","openai","cohere")}.}

\item{hard}{a \code{logical} specifying whether tokens should be overwritten.}
}
\value{
The function returns a \code{matrix} containing the similarities of all pairs of the embedding vectors. The \code{matrix} has \code{nrow(embedding)} rows and columns.
}
\description{
\code{set_tokens} sets access tokens of embedding APIs.
}
\details{
To obtain API tokens users must register with the respective services:

\itemize{
 \item{"Hugging Face"}{Sign up at \href{https://huggingface.co/inference-api}{huggingface.co/inference-api}. Then go to profile and click on settings. Generate access token in the \emph{Access Tokens} sub-menu.}
 \item{"OpenAI"}{Sign up at \href{https://openai.com/blog/openai-api}{openai.com/blog/openai-api}. Select API. Click on Personal (top-right) and select \emph{View API keys}. Then \emph{Create new secret key}.}
 \item{"Cohere"}{Sign up at \href{https://cohere.com/}{cohere.com}. Select \emph{API keys} in your personal dashboard. Either \emph{Create Trial key} (free of charge, but rate limited) or \emph{Create Production key} (costly, but unlimited). }
}

\emph{Hugging Face} and \emph{Cohere} both offer free-of-charge rate-limited API use.
}
\examples{

# get hugging face token
set_tokens("huggingface" = "TOKEN")

}
\references{
Wulff, D. U., Aeschbach, S., & Mata, R. (2024). embeddeR. psyArXiv
}
