% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{url.exists}
\alias{url.exists}
\title{Check if URL exists}
\usage{
url.exists(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
\code{TRUE} if the URL responds without error, otherwise \code{FALSE}.
}
\description{
Given a character string, returns a logical vector indicating
whether a request for a specific URL responds without error.
}
