% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{frame}
\alias{frame}
\title{Create data frame}
\usage{
frame(projection)
}
\arguments{
\item{projection}{a \code{numeric} matrix containing a text embedding projection generated with \code{k = 2}.}
}
\value{
The function returns a \code{tibble} containing a projected embedding. The projection dimensions will be assigned to columns \code{x} and \code{y}.
}
\description{
\code{frame} transforms an embedding projections to a data frame, specifically, a \code{tibble}.
}
\examples{

# embed, project, to tibble
projection <- embed(neo$text) \%>\%
  project() \%>\%
  to_tibble()

}
\references{
Wulff, D. U., Aeschbach, S., & Mata, R. (2024). embeddeR. psyArXiv
}
