% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.embedR_tbl}
\alias{plot.embedR_tbl}
\title{Plot embedding}
\usage{
\method{plot}{embedR_tbl}(
  data,
  dimensions = c(1, 2),
  label = "text",
  color = NULL,
  size = NULL,
  pt_padding = 1,
  box_padding = 1,
  label_size = 3,
  label_filter = NULL
)
}
\arguments{
\item{data}{a \code{data.frame} created using \link[emebedR]{frame}.}

\item{dimensions}{a \code{integer} vector determining the dimensions plotted. Default is \code{1:2}.}
}
\value{
The function returns a \code{matrix} containing projected coordinates for each embedding vectors. The \code{matrix} has \code{nrow(embedding)} rows and \code{k} columns.
}
\description{
\code{project} projects the embedding into a two or more dimensional space.
}
\examples{

# get embedding
embedding <- embed(neo$text)

# project embedding
project(embedding)

}
\references{
Wulff, D. U., Aeschbach, S., & Mata, R. (2024). embeddeR. psyArXiv
}
