% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.embedR}
\alias{print.embedR}
\title{Generic print for class mt_heatmap_raw}
\usage{
\method{print}{embedR}(x, n = 5, m = 5, ...)
}
\arguments{
\item{x}{an object of class embedR}

\item{n}{an \code{integer} specifying the number of rows to print. Default is \code{5}.}

\item{m}{an \code{integer} specifying the number of columns to print. Default is \code{5}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{print.embedR} abbreviates the \code{matrix} \code{print} and shows the attributes in a \code{tibble}.
}
